% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{availablePlotColumns}
\alias{availablePlotColumns}
\title{Available columns to use in \code{facet} and \code{colour} arguments in plot functions.}
\usage{
availablePlotColumns(result)
}
\arguments{
\item{result}{A summarised_result object.}
}
\value{
Character vector with the available columns.
}
\description{
Available columns to use in \code{facet} and \code{colour} arguments in plot functions.
}
\examples{
{
cdm <- mockCohortCharacteristics()

result <- summariseCharacteristics(cdm$cohort1)

availablePlotColumns(result)

mockDisconnect(cdm)
}
}
