% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{availableTableColumns}
\alias{availableTableColumns}
\title{Available columns to use in \code{header}, \code{groupColumn} and \code{hide} arguments in
table functions.}
\usage{
availableTableColumns(result)
}
\arguments{
\item{result}{A summarised_result object.}
}
\value{
Character vector with the available columns.
}
\description{
Available columns to use in \code{header}, \code{groupColumn} and \code{hide} arguments in
table functions.
}
\examples{
{
cdm <- mockCohortCharacteristics()

result <- summariseCharacteristics(cdm$cohort1)

availableTableColumns(result)

mockDisconnect(cdm)
}
}
