% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-package.R
\docType{package}
\name{CohortCharacteristics-package}
\alias{CohortCharacteristics}
\alias{CohortCharacteristics-package}
\title{CohortCharacteristics: Summarise and Visualise Characteristics of Patients in the OMOP CDM}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Summarise and visualise the characteristics of patients in data mapped to the Observational Medical Outcomes Partnership (OMOP) common data model (CDM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu-dev.github.io/CohortCharacteristics/}
  \item Report bugs at \url{https://github.com/darwin-eu-dev/CohortCharacteristics/issues}
}

}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

Authors:
\itemize{
  \item Yuchen Guo \email{yuchen.guo@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-0847-4855}{ORCID})
  \item Kim Lopez-Guell \email{kim.lopez@spc.ox.ac.uk} (\href{https://orcid.org/0000-0002-8462-8668}{ORCID})
  \item Edward Burn \email{edward.burn@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9286-1128}{ORCID})
  \item Nuria Mercade-Besora \email{nuria.mercadebesora@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0006-7948-3747}{ORCID})
  \item Marta Alcalde \email{marta.alcaldeherraiz@ndorms.ox.ac.uk} (\href{https://orcid.org/0009-0002-4405-1814}{ORCID})
}

Other contributors:
\itemize{
  \item Mike Du \email{mike.du@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0002-9517-8834}{ORCID}) [contributor]
}

}
\keyword{internal}
