% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortCount.R
\name{summariseCohortCount}
\alias{summariseCohortCount}
\title{Summarise cohort counts of a cohort_table}
\usage{
summariseCohortCount(cohort, cohortId = NULL)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{cohortId}{A cohort_definition_id to restrict, if NULL, all cohorts are
used.}
}
\value{
A summarised_result_object with the cohort counts summarised.
}
\description{
Summarise cohort counts of a cohort_table
}
