% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCohortDetails.R
\name{summariseCohortDetail}
\alias{summariseCohortDetail}
\title{Summarise attrition of a cohort_table}
\usage{
summariseCohortDetail(cohort, cohortId = NULL)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{cohortId}{A cohort_definition_id to restrict, if NULL, all cohorts are
used.}
}
\value{
A summarised_result_object with the attrition summarised.
}
\description{
Summarise attrition of a cohort_table
}
