% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortAttrition.R
\name{tableCohortAttrition}
\alias{tableCohortAttrition}
\title{Create a visual table from the output of summariseCohortAttrition.
`r lifecycle::badge("experimental")`}
\usage{
tableCohortAttrition(
  result,
  header = "cdm_name",
  groupColumn = "cohort_name",
  type = "gt"
)
}
\arguments{
\item{result}{A summarised_result object generated by summariseCohortAttrition().}

\item{header}{Columns to use as headers.}

\item{groupColumn}{Columns to use to group.}

\item{type}{Whether a 'gt', 'flextable' or a 'tibble' should be created.}
}
\value{
A visual table.
}
\description{
Create a visual table from the output of summariseCohortAttrition.
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

cdm$cohort2 |>
  summariseCohortAttrition() |>
  tableCohortAttrition()
}

}
