% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCharacteristics.R
\name{plotCharacteristics}
\alias{plotCharacteristics}
\title{Create a ggplot from the output of summariseCharacteristics.}
\usage{
plotCharacteristics(result, plotStyle = "barplot", facet = NULL, colour = NULL)
}
\arguments{
\item{result}{A summariseCharacteristics result.}

\item{plotStyle}{Either \code{barplot}, \code{scatterplot} or \code{boxplot}. If \code{barplot}
or \code{scatterplot} subset to just one estimate.}

\item{facet}{Columns to facet by. See options with \code{tidyColumns(result)}.
Formula is also allowed to specify rows and columns.}

\item{colour}{Columns to color by. See options with \code{tidyColumns(result)}.}
}
\value{
A ggplot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\donttest{
library(CohortCharacteristics)
library(dplyr)

cdm <- mockCohortCharacteristics()

results <- summariseCharacteristics(
  cohort = cdm$cohort1,
  ageGroup = list(c(0, 19), c(20, 39), c(40, 59), c(60, 79), c(80, 150)),
  tableIntersectCount = list(
    tableName = "visit_occurrence", window = c(-365, -1)
  ),
  cohortIntersectFlag = list(
    targetCohortTable = "cohort2", window = c(-365, -1)
  )
)

results |>
  filter(
    variable_name == "Cohort2 flag -365 to -1", estimate_name == "percentage"
  ) |>
  plotCharacteristics(
    plotStyle = "barplot",
    colour = "variable_level",
    facet = c("cdm_name", "cohort_name")
  )

results |>
  filter(variable_name == "Age", estimate_name == "mean") |>
  plotCharacteristics(
    plotStyle = "scatterplot",
    facet = "cdm_name"
  )

results |>
  filter(variable_name == "Age") |>
  plotCharacteristics(
    plotStyle = "boxplot",
    facet = "cdm_name",
    colour = "cohort_name"
  )

mockDisconnect(cdm = cdm)
}
}
