% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortAttrition.R
\name{plotCohortAttrition}
\alias{plotCohortAttrition}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotCohortAttrition(x)
}
\arguments{
\item{x}{attrition table}
}
\value{
A dgr_graph
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CDMConnector)
library(DrugUtilisation)
library(dplyr)
library(DiagrammeR)
library(PatientProfiles)

cdm <- mockDrugUtilisation(n = 1000)

cdm[["cohort1"]] <- cdm[["cohort1"]] |>
 filter(year(cohort_start_date) >= 2000) |>
  recordCohortAttrition("Restrict to cohort_start_date >= 2000") |>
  filter(year(cohort_end_date) < 2020) |>
  recordCohortAttrition("Restrict to cohort_end_date < 2020") |>
  compute(temporary = FALSE, name = "cohort1")

ca <- attrition(cdm[["cohort1"]]) |>
  filter(cohort_definition_id == 2)

render_graph(plotCohortAttrition(ca))
}
}
