% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortAttrition.R
\name{plotCohortAttrition}
\alias{plotCohortAttrition}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotCohortAttrition(result, cohortId = lifecycle::deprecated())
}
\arguments{
\item{result}{A summarised_result object with the results of
summariseCohortAttrition().}

\item{cohortId}{deprecated.}
}
\value{
A `grViz` visualisation.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(omopgenerics)
library(dplyr)

cdm <- mockCohortCharacteristics(numberIndividuals = 1000)

cdm[["cohort1"]] <- cdm[["cohort1"]] |>
  filter(year(cohort_start_date) >= 2000) |>
  recordCohortAttrition("Restrict to cohort_start_date >= 2000") |>
  filter(year(cohort_end_date) < 2020) |>
  recordCohortAttrition("Restrict to cohort_end_date < 2020") |>
  compute(temporary = FALSE, name = "cohort1")

cdm$cohort1 |>
  summariseCohortAttrition() |>
  filter(group_level == "cohort_2") |>
  plotCohortAttrition(cohortId = 2)
}

}
