% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortAttrition.R
\name{plotCohortAttrition}
\alias{plotCohortAttrition}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotCohortAttrition(x)
}
\arguments{
\item{x}{attrition table}
}
\value{
A dgr_graph
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CDMConnector)
library(CodelistGenerator)
library(dplyr)
library(DiagrammeR)
con <- DBI::dbConnect(duckdb::duckdb(),
dbdir = CDMConnector::eunomia_dir())
cdm <- CDMConnector::cdm_from_con(con,
cdm_schem = "main",
write_schema = "main")
meds_cs <- getDrugIngredientCodes(cdm = cdm, name = c("acetaminophen",
"morphine"))
cdm <- generateConceptCohortSet(
cdm = cdm,
name = "meds",
conceptSet = meds_cs,
end = "event_end_date",
limit = "all",
overwrite = TRUE)

cdm[["meds"]] <- cdm[["meds"]] |>
filter(year(cohort_end_date) <= 1949) |>
recordCohortAttrition("Restrict to cohort_start_date <= 1949") |>
filter(year(cohort_start_date) >= 1920) |>
recordCohortAttrition("Restrict to cohort_end_date >= 1920") |>
compute(temporary = FALSE, name = "meds")

ca <- attrition(cdm[["meds"]]) |> filter(cohort_definition_id == 2) |>
mutate(number_records = 161137831)

render_graph(plotCohortAttrition(ca))
}
}
