% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortCount.R
\name{plotCohortCount}
\alias{plotCohortCount}
\title{Plot the result of summariseCohortCount.}
\usage{
plotCohortCount(result, facet = c("cdm_name"), colour = NULL, x = NULL)
}
\arguments{
\item{result}{A summarised_result object. Output of summariseCohortCount().}

\item{facet}{Variables to facet by.}

\item{colour}{Variables to colour by.}

\item{x}{Variables to use in x axis.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)
library(PatientProfiles)
library(dplyr)

cdm <- mockCohortCharacteristics(numberIndividuals = 100)
counts <- cdm$cohort2 |>
  addSex() |>
  addAge(ageGroup = list(c(0, 29), c(30, 59), c(60, Inf))) |>
  summariseCohortCount(strata = list("age_group", "sex", c("age_group", "sex"))) |>
  filter(variable_name == "Number subjects")
counts |>
  plotCohortCount(
    facet = cohort_name ~ age_group,
    colour = "sex",
    x = "sex")
}

}
