% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortOverlap.R
\name{plotCohortOverlap}
\alias{plotCohortOverlap}
\title{Plot the result of summariseCohortOverlap.}
\usage{
plotCohortOverlap(
  result,
  facet = c("cdm_name", "cohort_name_reference"),
  uniqueCombinations = FALSE,
  y = NULL,
  .options = lifecycle::deprecated()
)
}
\arguments{
\item{result}{A summarised_result object. Output of summariseCohortOverlap().}

\item{facet}{Columns to facet by. See options with `tidyColumns(result)`.
Formula is also allowed to specify rows and columns.}

\item{uniqueCombinations}{Whether to restrict to unique reference and
comparator comparisons.}

\item{y}{Variables to use in y axis, if NULL all variables not present in
facet are used.}

\item{.options}{deprecated.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)
cdm <- CohortCharacteristics::mockCohortCharacteristics()
overlap <- summariseCohortOverlap(cdm$cohort2)
plotCohortOverlap(overlap)
}

}
