% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortTiming.R
\name{plotCohortTiming}
\alias{plotCohortTiming}
\title{Plot summariseCohortTiming results.}
\usage{
plotCohortTiming(
  result,
  plotType = "boxplot",
  timeScale = "days",
  facetVarX = "variable_name",
  facetVarY = "group_level",
  colour = NULL,
  colour_name = NULL,
  uniqueCombinations = TRUE
)
}
\arguments{
\item{result}{A summariseCohortTiming result.}

\item{plotType}{Can eiter be boxplot or density.}

\item{timeScale}{Time scale to plot results. Can be days or years.}

\item{facetVarX}{column in data to facet by on horizontal axis}

\item{facetVarY}{column in data to facet by on vertical axis}

\item{colour}{Variables to use for colours}

\item{colour_name}{Colour legend name}

\item{uniqueCombinations}{If TRUE, only unique combinations of reference and
comparator plots will be plotted.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)
cdm <- CohortCharacteristics::mockCohortCharacteristics()
timing <- summariseCohortTiming(cdm$cohort2)
plotCohortTiming(timing)
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
