% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCohortTiming.R
\name{plotCohortTiming}
\alias{plotCohortTiming}
\title{Plot summariseCohortTiming results.}
\usage{
plotCohortTiming(
  result,
  plotType = "boxplot",
  timeScale = "days",
  facet = c("cdm_name", "cohort_name_reference"),
  colour = c("cohort_name_comparator"),
  colourName = lifecycle::deprecated(),
  uniqueCombinations = FALSE,
  .options = lifecycle::deprecated()
)
}
\arguments{
\item{result}{A summarised_result object. Output of summariseCohortTiming().}

\item{plotType}{Type of desired formatted table, possibilities are "boxplot" and
"density".}

\item{timeScale}{Time scale to plot results. Can be days or years.}

\item{facet}{variables to facet by.}

\item{colour}{Variables to use for colours.}

\item{colourName}{deprecated.}

\item{uniqueCombinations}{If TRUE, only unique combinations of reference and
comparator plots will be plotted.}

\item{.options}{deprecated.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
