% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparedLargeScaleCharacteristics.R
\name{plotComparedLargeScaleCharacteristics}
\alias{plotComparedLargeScaleCharacteristics}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotComparedLargeScaleCharacteristics(
  result,
  reference,
  facet = NULL,
  colour = NULL,
  missings = 0
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseLargeScaleCharacteristics().}

\item{reference}{A named character to set up the reference.}

\item{facet}{Variables to facet by. Use tidyColumns to see options.}

\item{colour}{column in data to color by. Use tidyColumns to see options.}

\item{missings}{value to replace the missings with.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
