% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparedLargeScaleCharacteristics.R
\name{plotComparedLargeScaleCharacteristics}
\alias{plotComparedLargeScaleCharacteristics}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotComparedLargeScaleCharacteristics(
  data,
  referenceGroupLevel = NULL,
  referenceStrataLevel = NULL,
  referenceVariableLevel = NULL,
  referenceCdmName = NULL,
  splitStrata = FALSE,
  facet = NULL,
  colorVars = NULL,
  missings = 0
)
}
\arguments{
\item{data}{output of summariseLargeScaleCharacteristics().}

\item{referenceGroupLevel}{group_level value to be used as the reference.}

\item{referenceStrataLevel}{strata_level value to be used as the reference.}

\item{referenceVariableLevel}{variable_level value to be used as the reference.}

\item{referenceCdmName}{cdm_name value to be used as the reference.}

\item{splitStrata}{boolean variable (TRUE/FALSE)}

\item{facet}{columns in data to facet. If the facet position wants to be specified, use the formula class for the input
(e.g., strata + table_name ~ group_level + cdm_name). Variables before "~" will be facet by on horizontal axis, whereas those after "~" on vertical axis.
Character format is also allowed (e.g., c("strata","table_name","group_level","cdm_name")).
Only the following columns are allowed to be facet by: c("cdm_name", "group_level", "strata_level", "variable_level", "strata", "table_name").
If splitStrata = TRUE, strata levels are also allowed.}

\item{colorVars}{column in data to color by. Only the following columns are allowed to be used: c("cdm_name", "group_level", "strata_level", "variable_level", "strata", "table_name").
If splitStrata = TRUE, strata levels are also allowed.}

\item{missings}{value to replace the missings with.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)
library(DrugUtilisation)
cdm <- DrugUtilisation::mockDrugUtilisation()

lsc <- CohortCharacteristics::summariseLargeScaleCharacteristics(cdm$cohort1,
eventInWindow = "condition_occurrence", episodeInWindow = "drug_exposure",
minimumFrequency = 0.05)

plotComparedLargeScaleCharacteristics(data = lsc,
referenceGroupLevel = "cohort_2", referenceStrataLevel = NULL,
referenceVariableLevel = '-inf to -366', referenceCdmName = NULL,
splitStrata = TRUE, facet = variable_level ~ group_level, colorVars = NULL,
missings = 0)
CDMConnector::cdmDisconnect(cdm = cdm)
}
}
