% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComparedLargeScaleCharacteristics.R
\name{plotComparedLargeScaleCharacteristics}
\alias{plotComparedLargeScaleCharacteristics}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotComparedLargeScaleCharacteristics(
  result,
  reference,
  facet = NULL,
  colour = NULL,
  missings = 0
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseLargeScaleCharacteristics().}

\item{reference}{A named character to set up the reference.}

\item{facet}{Columns to facet by. See options with `tidyColumns(result)`.
Formula is also allowed to specify rows and columns.}

\item{colour}{Columns to color by. See options with `tidyColumns(result)`.}

\item{missings}{value to replace the missings with.}
}
\value{
A ggplot.
}
\description{
`r lifecycle::badge("experimental")`
}
