% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLargeScaleCharacteristics.R
\name{plotLargeScaleCharacteristics}
\alias{plotLargeScaleCharacteristics}
\title{create a ggplot from the output of summariseLargeScaleCharacteristics.}
\usage{
plotLargeScaleCharacteristics(
  result,
  facet = c("cdm_name", "cohort_name"),
  colour = "variable_level"
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseLargeScaleCharacteristics().}

\item{facet}{Character vector that indicates the columns to facet by, you can
use any tidyColumns(result). Formula als is allowed to specify rows and
columns.}

\item{colour}{Character vector that indicates the columns to colour by, you
can use any tidyColumns(result).}
}
\value{
A ggplot2 object.
}
\description{
`r lifecycle::badge("experimental")`
}
