% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCharacteristics.R
\name{tableCharacteristics}
\alias{tableCharacteristics}
\title{Format a summarise_characteristics object into a visual table.}
\usage{
tableCharacteristics(
  result,
  type = "gt",
  header = c("group"),
  groupColumn = NULL,
  hide = c("result_id", "estimate_type", "additional_name", "additional_level"),
  .options = list(),
  formatEstimateName = lifecycle::deprecated(),
  split = lifecycle::deprecated(),
  excludeColumns = lifecycle::deprecated()
)
}
\arguments{
\item{result}{A summarise_characteristics object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{header}{A vector containing which elements should go into the header
in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`,
`variable`, `estimate`, `settings`.}

\item{groupColumn}{Column to use as group labels.}

\item{hide}{Columns to drop from the output table.}

\item{.options}{Named list with additional formatting options.
CohortCharacteristics::optionsTableCharacteristics() shows allowed arguments and
their default values.}

\item{formatEstimateName}{deprecated.}

\item{split}{deprecated.}

\item{excludeColumns}{deprecated.}
}
\value{
A table with a formatted version of the summariseCharacteristics
result.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
cdm <- mockCohortCharacteristics()

cdm$cohort1 |>
  summariseCharacteristics() |>
  tableCharacteristics()

mockDisconnect(cdm = cdm)
}

}
