% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCharacteristics.R
\name{tableCharacteristics}
\alias{tableCharacteristics}
\title{Format a summarise_characteristics object into a visual table.}
\usage{
tableCharacteristics(
  result,
  type = "gt",
  header = c("cdm_name", "cohort_name"),
  groupColumn = NULL
)
}
\arguments{
\item{result}{A summarise_characteristics object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{header}{A vector containing which elements should go into the header
in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`,
`variable`, `estimate`, `settings`.}

\item{groupColumn}{Column to use as group labels.}
}
\value{
A table with a formatted version of the summariseCharacteristics
result.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
cdm <- mockCohortCharacteristics()

result <- summariseCharacteristics(cdm$cohort1)

tableCharacteristics(result)

mockDisconnect(cdm = cdm)
}

}
