% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortAttrition.R
\name{tableCohortAttrition}
\alias{tableCohortAttrition}
\title{Create a visual table from the output of summariseCohortAttrition.
`r lifecycle::badge("experimental")`}
\usage{
tableCohortAttrition(
  result,
  type = "gt",
  header = "variable_name",
  groupColumn = c("cdm_name", "cohort_name")
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseCohortAttrition().}

\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as headers.}

\item{groupColumn}{Columns to use to group.}
}
\value{
A visual table.
}
\description{
Create a visual table from the output of summariseCohortAttrition.
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

result <- summariseCohortAttrition(cdm$cohort2)

tableCohortAttrition(result)

mockDisconnect(cdm)
}

}
