% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortAttrition.R
\name{tableCohortAttrition}
\alias{tableCohortAttrition}
\title{Create a visual table from the output of summariseCohortAttrition.}
\usage{
tableCohortAttrition(
  result,
  type = "gt",
  header = "variable_name",
  groupColumn = c("cdm_name", "cohort_name")
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseCohortAttrition().}

\item{type}{Type of table. Check supported types with
\code{visOmopResults::tableType()}.}

\item{header}{Columns to use as header. See options with
\code{tidyColumns(result)}.}

\item{groupColumn}{Columns to group by. See options with
\code{tidyColumns(result)}.}
}
\value{
A visual table.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

result <- summariseCohortAttrition(cdm$cohort2)

tableCohortAttrition(result)

mockDisconnect(cdm)
}

}
