% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortCount.R
\name{tableCohortCount}
\alias{tableCohortCount}
\title{Format a summarise_characteristics object into a visual table.}
\usage{
tableCohortCount(
  result,
  type = "gt",
  header = "cohort_name",
  groupColumn = NULL,
  hide = "variable_level"
)
}
\arguments{
\item{result}{A summarise_characteristics object.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{header}{A vector containing which elements should go into the header
in order. Allowed are: `cdm_name`, `group`, `strata`, `additional`,
`variable`, `estimate`, `settings`.}

\item{groupColumn}{Column to use as group labels.}

\item{hide}{Columns to drop from the output table.}
}
\value{
A table with a formatted version of the summariseCohortCount result
result.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

result <- summariseCohortCount(cdm$cohort1)

tableCohortCount(result)

mockDisconnect(cdm = cdm)
}

}
