% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortCount.R
\name{tableCohortCount}
\alias{tableCohortCount}
\title{Create a visual table from the output of summariseCohortCount.
`r lifecycle::badge("experimental")`}
\usage{
tableCohortCount(
  result,
  header = "cdm_name",
  groupColumn = "cohort_name",
  type = "gt"
)
}
\arguments{
\item{result}{A summarised_result object generated by summariseCohortCount().}

\item{header}{Columns to use as headers.}

\item{groupColumn}{Columns to use to group.}

\item{type}{Whether a 'gt' or 'flextable' should be created.}
}
\value{
A visual table.
}
\description{
Create a visual table from the output of summariseCohortCount.
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

cdm$cohort2 |>
  summariseCohortCount() |>
  tableCohortCount()
}

}
