% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortCount.R
\name{tableCohortCount}
\alias{tableCohortCount}
\title{Format a summarise_characteristics object into a visual table.}
\usage{
tableCohortCount(
  result,
  type = "gt",
  header = "cohort_name",
  groupColumn = NULL
)
}
\arguments{
\item{result}{A summarised_result object. Output of summariseCohortCount().}

\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as header. See options with
`tidyColumns(result)`.}

\item{groupColumn}{Columns to group by. See options with
`tidyColumns(result)`.}
}
\value{
A table with a formatted version of the summariseCohortCount result
result.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

result <- summariseCohortCount(cdm$cohort1)

tableCohortCount(result)

mockDisconnect(cdm = cdm)
}

\donttest{
#' library(CohortCharacteristics)
library(dplyr, warn.conflicts = FALSE)

cdm <- mockCohortCharacteristics()

result <- summariseCohortCount(cohort = cdm$cohort1)

tableCohortCount(result)

mockDisconnect(cdm)
}


}
