% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortOverlap.R
\name{tableCohortOverlap}
\alias{tableCohortOverlap}
\title{Format a summariseOverlapCohort result into a visual table.}
\usage{
tableCohortOverlap(
  result,
  type = "gt",
  header = c("variable_name"),
  groupColumn = c("cdm_name"),
  hide = c("variable_level"),
  uniqueCombinations = TRUE
)
}
\arguments{
\item{result}{A summariseOverlapCohort result.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{header}{A vector containing which elements should go into the header
in order. Allowed are: \code{cdm_name}, \code{group}, \code{strata}, \code{additional},
\code{variable}, \code{estimate}, \code{settings}.}

\item{groupColumn}{Column to use as group labels.}

\item{hide}{Columns to drop from the output table.}

\item{uniqueCombinations}{Whether to display unique combinations
reference - comparator.}
}
\value{
A formatted table of the summariseOverlapCohort result.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\donttest{
library(CohortCharacteristics)
cdm <- mockCohortCharacteristics()
overlap <- summariseCohortOverlap(cdm$cohort2)
tableCohortOverlap(overlap)
mockDisconnect(cdm = cdm)
}

}
