% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortOverlap.R
\name{tableCohortOverlap}
\alias{tableCohortOverlap}
\title{Format a summariseOverlapCohort result into a visual table.}
\usage{
tableCohortOverlap(
  result,
  uniqueCombinations = TRUE,
  type = "gt",
  header = c("variable_name"),
  groupColumn = c("cdm_name")
)
}
\arguments{
\item{result}{A summarised_result object. Output of summariseCohortOverlap().}

\item{uniqueCombinations}{Whether to restrict to unique reference and
comparator comparisons.}

\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as header. See options with
`tidyColumns(result)`.}

\item{groupColumn}{Columns to group by. See options with
`tidyColumns(result)`.}
}
\value{
A formatted table of the summariseOverlapCohort result.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\donttest{
library(CohortCharacteristics)

cdm <- mockCohortCharacteristics()

overlap <- summariseCohortOverlap(cdm$cohort2)

tableCohortOverlap(overlap)

mockDisconnect(cdm)
}

}
