% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortTiming.R
\name{tableCohortTiming}
\alias{tableCohortTiming}
\title{Format a summariseCohortTiming result into a visual table.}
\usage{
tableCohortTiming(
  result,
  timeScale = "days",
  type = "gt",
  header = visOmopResults::strataColumns(result),
  groupColumn = NULL,
  hide = NULL,
  uniqueCombinations = TRUE
)
}
\arguments{
\item{result}{A summariseCohortTiming result}

\item{timeScale}{Time scale to plot results. Can be days or years.}

\item{type}{Type of desired formatted table, possibilities: "gt",
"flextable", "tibble".}

\item{header}{Columns to use as header. See options with tidyColumns(result).}

\item{groupColumn}{Columns to use as group labels. See options with
tidyColumns(result).}

\item{hide}{Columns to hide. See options with tidyColumns(result).}

\item{uniqueCombinations}{Whether to restrict to unique reference and
comparator.}
}
\value{
A formatted table of the summariseCohortTiming result.
}
\description{
`r lifecycle::badge("experimental")`
}
