% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCohortTiming.R
\name{tableCohortTiming}
\alias{tableCohortTiming}
\title{Format a summariseCohortTiming result into a visual table.}
\usage{
tableCohortTiming(
  result,
  timeScale = "days",
  uniqueCombinations = TRUE,
  type = "gt",
  header = visOmopResults::strataColumns(result),
  groupColumn = NULL
)
}
\arguments{
\item{result}{A summarised_result object. Output of summariseCohortTiming().}

\item{timeScale}{Time scale to plot results. Can be days or years.}

\item{uniqueCombinations}{Whether to restrict to unique reference and
comparator.}

\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as header. See options with tidyColumns(result).}

\item{groupColumn}{Columns to use as group labels. See options with
tidyColumns(result).}
}
\value{
A formatted table of the summariseCohortTiming result.
}
\description{
`r lifecycle::badge("experimental")`
}
