% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableLargeScaleCharacteristics.R
\name{tableLargeScaleCharacteristics}
\alias{tableLargeScaleCharacteristics}
\title{Format a summarise_large_scale_characteristics object into a visual table.}
\usage{
tableLargeScaleCharacteristics(
  result,
  topConcepts = NULL,
  type = "gt",
  header = c("cdm_name", "cohort_name", visOmopResults::strataColumns(result),
    "window_name"),
  groupColumn = c("table_name", "type", "analysis")
)
}
\arguments{
\item{result}{A summarised_result object. Output of
summariseLargeScaleCharacteristics().}

\item{topConcepts}{Number of concepts to restrict the table.}

\item{type}{Type of table. Check supported types with
`visOmopResults::tableType()`.}

\item{header}{Columns to use as header. See options with
`tidyColumns(result)`.}

\item{groupColumn}{Columns to group by. See options with
`tidyColumns(result)`.}
}
\value{
A formatted table.
}
\description{
`r lifecycle::badge("experimental")`
}
\examples{
\dontrun{
library(DBI)
library(duckdb)
library(CDMConnector)

con <- dbConnect(duckdb(), eunomia_dir())
cdm <- cdmFromCon(con = con, cdmSchema = "main", writeSchema = "main")
cdm <- generateConceptCohortSet(
  cdm = cdm,
  conceptSet = list("viral_pharyngitis" = 4112343),
  name = "my_cohort"
)

result <- summariseLargeScaleCharacteristics(
  cohort = cdm$my_cohort,
  eventInWindow = "condition_occurrence",
  episodeInWindow = "drug_exposure"
)

tableLargeScaleCharacteristics(result)

cdmDisconnect(cdm)
}

}
